Creating a file for OptionProviderMixin and helper functions
